feat: Rename SparkLogger to SparkExceptionHandler
`SparkExceptionHandler` better reflect its purpose of handling exceptions within Spark
